WebVTT Cleanup (#1059)
* refactor vtt-segment-loader to extend segment-loader

* refactor segment loading tests into common shared tests and non shared
