+++ src/vtt-segment-loader.js
@@ -1,31 +1,13 @@
 /**
  * @file vtt-segment-loader.js
  */
-import {getMediaInfoForTime_ as getMediaInfoForTime} from './playlist';
+import SegmentLoader from './segment-loader';
 import videojs from 'video.js';
-import Config from './config';
 import window from 'global/window';
 import removeCuesFromTrack from 'videojs-contrib-media-sources/es5/remove-cues-from-track.js';
-import {mediaSegmentRequest, REQUEST_ERRORS} from './media-segment-request';
+import { initSegmentId } from './bin-utils';
 
-// in ms
-const CHECK_BUFFER_DELAY = 500;
 const VTT_LINE_TERMINATORS = new Uint8Array('\n\n'.split('').map(char => char.charCodeAt(0)));
-
-/**
- * Returns a unique string identifier for a media initialization
- * segment.
- */
-const initSegmentId = function(initSegment) {
-  let byterange = initSegment.byterange || {
-    length: Infinity,
-    offset: 0
-  };
-
-  return [
-    byterange.length, byterange.offset, initSegment.resolvedUri
-  ].join(',');
-};
 
 const uintToString = function(uintArray) {
   return String.fromCharCode.apply(null, uintArray);
@@ -38,149 +20,24 @@ const uintToString = function(uintArray) {
  * @param {Object} options required and optional options
  * @extends videojs.EventTarget
  */
-export default class VTTSegmentLoader extends videojs.EventTarget {
+export default class VTTSegmentLoader extends SegmentLoader {
   constructor(options) {
-    super();
-    // check pre-conditions
-    if (!options) {
-      throw new TypeError('Initialization options are required');
-    }
-    if (typeof options.currentTime !== 'function') {
-      throw new TypeError('No currentTime getter specified');
-    }
-    let settings = videojs.mergeOptions(videojs.options.hls, options);
+    super(options);
 
-    // public properties
-    this.state = 'INIT';
-    this.bandwidth = settings.bandwidth;
-    this.throughput = {rate: 0, count: 0};
-    this.roundTrip = NaN;
-    this.resetStats_();
-    this.mediaIndex = null;
-
-    // private settings
-    this.hasPlayed_ = settings.hasPlayed;
-    this.currentTime_ = settings.currentTime;
-    this.seekable_ = settings.seekable;
-    this.seeking_ = settings.seeking;
-    this.setCurrentTime_ = settings.setCurrentTime;
-    this.mediaSource_ = settings.mediaSource;
-    this.hls_ = settings.hls;
-    this.loaderType_ = settings.loaderType;
-
-    // private instance variables
-    this.checkBufferTimeout_ = null;
-    this.error_ = void 0;
-    this.currentTimeline_ = -1;
-    this.pendingSegment_ = null;
-    this.sourceUpdater_ = null;
-    this.xhrOptions_ = null;
+    // SegmentLoader requires a MediaSource be specified or it will throw an error;
+    // however, VTTSegmentLoader has no need of a media source, so delete the reference
+    this.mediaSource_ = null;
 
     this.subtitlesTrack_ = null;
-
-    // Fragmented mp4 playback
-    this.initSegments_ = {};
-
-    this.decrypter_ = settings.decrypter;
-
-    // Manages the tracking and generation of sync-points, mappings
-    // between a time in the display time and a segment index within
-    // a playlist
-    this.syncController_ = settings.syncController;
-    this.syncPoint_ = {
-      segmentIndex: 0,
-      time: 0
-    };
-
-    this.syncController_.on('syncinfoupdate', () => this.trigger('syncinfoupdate'));
-
-    // ...for determining the fetch location
-    this.fetchAtBuffer_ = false;
-
-    if (settings.debug) {
-      this.logger_ = videojs.log.bind(videojs, 'segment-loader', this.loaderType_, '->');
-    }
-  }
-
-  /**
-   * reset all of our media stats
-   *
-   * @private
-   */
-  resetStats_() {
-    this.mediaBytesTransferred = 0;
-    this.mediaRequests = 0;
-    this.mediaRequestsAborted = 0;
-    this.mediaRequestsTimedout = 0;
-    this.mediaRequestsErrored = 0;
-    this.mediaTransferDuration = 0;
-    this.mediaSecondsLoaded = 0;
-  }
-
-  /**
-   * dispose of the SegmentLoader and reset to the default state
-   */
-  dispose() {
-    this.state = 'DISPOSED';
-    this.abort_();
-    this.resetStats_();
-  }
-
-  /**
-   * abort anything that is currently doing on with the SegmentLoader
-   * and reset to a default state
-   */
-  abort() {
-    if (this.state !== 'WAITING') {
-      if (this.pendingSegment_) {
-        this.pendingSegment_ = null;
-      }
-      return;
-    }
-
-    this.abort_();
-
-    // don't wait for buffer check timeouts to begin fetching the
-    // next segment
-    if (!this.paused()) {
-      this.state = 'READY';
-      this.monitorBuffer_();
-    }
-  }
-
-  /**
-   * abort all pending xhr requests and null any pending segements
-   *
-   * @private
-   */
-  abort_() {
-    if (this.pendingSegment_) {
-      this.pendingSegment_.abortRequests();
-    }
-
-    // clear out the segment being processed
-    this.pendingSegment_ = null;
-  }
-
-  /**
-   * set an error on the segment loader and null out any pending segements
-   *
-   * @param {Error} error the error to set on the SegmentLoader
-   * @return {Error} the error that was set or that is currently set
-   */
-  error(error) {
-    if (typeof error !== 'undefined') {
-      this.error_ = error;
-    }
-
-    this.pendingSegment_ = null;
-    return this.error_;
   }
 
   /**
    * Indicates which time ranges are buffered
+   *
+   * @return {TimeRange}
+   *         TimeRange object representing the current buffered ranges
    */
-  buffered() {
+  buffered_() {
     if (!this.subtitlesTrack_ || !this.subtitlesTrack_.cues.length) {
       return videojs.createTimeRanges();
     }
@@ -232,34 +89,15 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
   }
 
   /**
-   * load a playlist and start to fill the buffer
+   * Returns true if all configuration required for loading is present, otherwise false.
+   *
+   * @return {Boolean} True if the all configuration is ready for loading
+   * @private
    */
-  load() {
-    // un-pause
-    this.monitorBuffer_();
-
-    // if we don't have a playlist yet, keep waiting for one to be
-    // specified
-    if (!this.playlist_) {
-      return;
-    }
-
-    // not sure if this is the best place for this
-    this.syncController_.setDateTimeMapping(this.playlist_);
-
-    // if all the configuration is ready, initialize and begin loading
-    if (this.state === 'INIT' && this.subtitlesTrack_) {
-      return this.init_();
-    }
-
-    // if we're in the middle of processing a segment already, don't
-    // kick off an additional segment request
-    if (this.state !== 'READY' &&
-        this.state !== 'INIT') {
-      return;
-    }
-
-    this.state = 'READY';
+  couldBeginLoading_() {
+    return this.playlist_ &&
+           this.subtitlesTrack_ &&
+           !this.paused();
   }
 
   /**
@@ -286,140 +124,9 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
 
     // if we were unpaused but waiting for a sourceUpdater, start
     // buffering now
-    if (this.playlist_ &&
-        this.state === 'INIT' &&
-        !this.paused() &&
-        this.subtitlesTrack_) {
+    if (this.state === 'INIT' && this.couldBeginLoading_()) {
       this.init_();
     }
-  }
-
-  /**
-   * set a playlist on the segment loader
-   *
-   * @param {PlaylistLoader} media the playlist to set on the segment loader
-   */
-  playlist(newPlaylist, options = {}) {
-    if (!newPlaylist) {
-      return;
-    }
-
-    let oldPlaylist = this.playlist_;
-    let segmentInfo = this.pendingSegment_;
-
-    this.playlist_ = newPlaylist;
-    this.xhrOptions_ = options;
-
-    // when we haven't started playing yet, the start of a live playlist
-    // is always our zero-time so force a sync update each time the playlist
-    // is refreshed from the server
-    if (!this.hasPlayed_()) {
-      newPlaylist.syncInfo = {
-        mediaSequence: newPlaylist.mediaSequence,
-        time: 0
-      };
-    }
-
-    // in VOD, this is always a rendition switch (or we updated our syncInfo above)
-    // in LIVE, we always want to update with new playlists (including refreshes)
-    this.trigger('syncinfoupdate');
-
-    // if we were unpaused but waiting for a playlist, start
-    // buffering now
-    if (this.subtitlesTrack_ && this.state === 'INIT' && !this.paused()) {
-      return this.init_();
-    }
-
-    if (!oldPlaylist || oldPlaylist.uri !== newPlaylist.uri) {
-      if (this.mediaIndex !== null) {
-        // we must "resync" the segment loader when we switch renditions and
-        // the segment loader is already synced to the previous rendition
-        this.resyncLoader();
-      }
-
-      // the rest of this function depends on `oldPlaylist` being defined
-      return;
-    }
-
-    // we reloaded the same playlist so we are in a live scenario
-    // and we will likely need to adjust the mediaIndex
-    let mediaSequenceDiff = newPlaylist.mediaSequence - oldPlaylist.mediaSequence;
-
-    this.logger_('mediaSequenceDiff', mediaSequenceDiff);
-
-    // update the mediaIndex on the SegmentLoader
-    // this is important because we can abort a request and this value must be
-    // equal to the last appended mediaIndex
-    if (this.mediaIndex !== null) {
-      this.mediaIndex -= mediaSequenceDiff;
-    }
-
-    // update the mediaIndex on the SegmentInfo object
-    // this is important because we will update this.mediaIndex with this value
-    // in `handleUpdateEnd_` after the segment has been successfully appended
-    if (segmentInfo) {
-      segmentInfo.mediaIndex -= mediaSequenceDiff;
-
-      // we need to update the referenced segment so that timing information is
-      // saved for the new playlist's segment, however, if the segment fell off the
-      // playlist, we can leave the old reference and just lose the timing info
-      if (segmentInfo.mediaIndex >= 0) {
-        segmentInfo.segment = newPlaylist.segments[segmentInfo.mediaIndex];
-      }
-    }
-
-    this.syncController_.saveExpiredSegmentInfo(oldPlaylist, newPlaylist);
-  }
-
-  /**
-   * Prevent the loader from fetching additional segments. If there
-   * is a segment request outstanding, it will finish processing
-   * before the loader halts. A segment loader can be unpaused by
-   * calling load().
-   */
-  pause() {
-    if (this.checkBufferTimeout_) {
-      window.clearTimeout(this.checkBufferTimeout_);
-
-      this.checkBufferTimeout_ = null;
-    }
-  }
-
-  /**
-   * Returns whether the segment loader is fetching additional
-   * segments when given the opportunity. This property can be
-   * modified through calls to pause() and load().
-   */
-  paused() {
-    return this.checkBufferTimeout_ === null;
-  }
-
-  /**
-   * Delete all the buffered data and reset the SegmentLoader
-   */
-  resetEverything() {
-    this.resetLoader();
-    this.remove(0, Infinity);
-  }
-
-  /**
-   * Force the SegmentLoader to resync and start loading around the currentTime instead
-   * of starting at the end of the buffer
-   *
-   * Useful for fast quality changes
-   */
-  resetLoader() {
-    this.fetchAtBuffer_ = false;
-    this.resyncLoader();
-  }
-
-  /**
-   * Force the SegmentLoader to restart synchronization and make a conservative guess
-   * before returning to the simple walk-forward method
-   */
-  resyncLoader() {
-    this.mediaIndex = null;
-    this.syncPoint_ = null;
   }
 
   /**
@@ -429,38 +136,6 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
    */
   remove(start, end) {
     removeCuesFromTrack(start, end, this.subtitlesTrack_);
-  }
-
-  /**
-   * (re-)schedule monitorBufferTick_ to run as soon as possible
-   *
-   * @private
-   */
-  monitorBuffer_() {
-    if (this.checkBufferTimeout_) {
-      window.clearTimeout(this.checkBufferTimeout_);
-    }
-
-    this.checkBufferTimeout_ = window.setTimeout(this.monitorBufferTick_.bind(this), 1);
-  }
-
-  /**
-   * As long as the SegmentLoader is in the READY state, periodically
-   * invoke fillBuffer_().
-   *
-   * @private
-   */
-  monitorBufferTick_() {
-    if (this.state === 'READY') {
-      this.fillBuffer_();
-    }
-
-    if (this.checkBufferTimeout_) {
-      window.clearTimeout(this.checkBufferTimeout_);
-    }
-
-    this.checkBufferTimeout_ = window.setTimeout(this.monitorBufferTick_.bind(this),
-                                                 CHECK_BUFFER_DELAY);
   }
 
   /**
@@ -475,13 +150,13 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
   fillBuffer_() {
     if (!this.syncPoint_) {
       this.syncPoint_ = this.syncController_.getSyncPoint(this.playlist_,
-                                                          this.mediaSource_.duration,
+                                                          this.duration_(),
                                                           this.currentTimeline_,
                                                           this.currentTime_());
     }
 
     // see if we need to begin loading immediately
-    let segmentInfo = this.checkBuffer_(this.buffered(),
+    let segmentInfo = this.checkBuffer_(this.buffered_(),
                                         this.playlist_,
                                         this.mediaIndex,
                                         this.hasPlayed_(),
@@ -509,169 +184,8 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
       this.state = 'WAITING_ON_TIMELINE';
       return;
     }
+
     this.loadSegment_(segmentInfo);
-  }
-
-  /**
-   * Determines what segment request should be made, given current playback
-   * state.
-   *
-   * @param {TimeRanges} buffered - the state of the buffer
-   * @param {Object} playlist - the playlist object to fetch segments from
-   * @param {Number} mediaIndex - the previous mediaIndex fetched or null
-   * @param {Boolean} hasPlayed - a flag indicating whether we have played or not
-   * @param {Number} currentTime - the playback position in seconds
-   * @param {Object} syncPoint - a segment info object that describes the
-   * @returns {Object} a segment request object that describes the segment to load
-   */
-  checkBuffer_(buffered, playlist, mediaIndex, hasPlayed, currentTime, syncPoint) {
-    let lastBufferedEnd = 0;
-    let startOfSegment;
-
-    if (buffered.length) {
-      lastBufferedEnd = buffered.end(buffered.length - 1);
-    }
-
-    let bufferedTime = Math.max(0, lastBufferedEnd - currentTime);
-
-    if (!playlist.segments.length) {
-      return null;
-    }
-
-    // if there is plenty of content buffered, and the video has
-    // been played before relax for awhile
-    if (bufferedTime >= Config.GOAL_BUFFER_LENGTH) {
-      return null;
-    }
-
-    // if the video has not yet played once, and we already have
-    // one segment downloaded do nothing
-    if (!hasPlayed && bufferedTime >= 1) {
-      return null;
-    }
-
-    this.logger_('checkBuffer_',
-      'mediaIndex:', mediaIndex,
-      'hasPlayed:', hasPlayed,
-      'currentTime:', currentTime,
-      'syncPoint:', syncPoint,
-      'fetchAtBuffer:', this.fetchAtBuffer_,
-      'bufferedTime:', bufferedTime);
-
-    // When the syncPoint is null, there is no way of determining a good
-    // conservative segment index to fetch from
-    // The best thing to do here is to get the kind of sync-point data by
-    // making a request
-    if (syncPoint === null) {
-      mediaIndex = this.getSyncSegmentCandidate_(playlist);
-      this.logger_('getSync', 'mediaIndex:', mediaIndex);
-      return this.generateSegmentInfo_(playlist, mediaIndex, null, true);
-    }
-
-    // Under normal playback conditions fetching is a simple walk forward
-    if (mediaIndex !== null) {
-      this.logger_('walkForward', 'mediaIndex:', mediaIndex + 1);
-      let segment = playlist.segments[mediaIndex];
-
-      if (segment && segment.end) {
-        startOfSegment = segment.end;
-      } else {
-        startOfSegment = lastBufferedEnd;
-      }
-      return this.generateSegmentInfo_(playlist, mediaIndex + 1, startOfSegment, false);
-    }
-
-    // There is a sync-point but the lack of a mediaIndex indicates that
-    // we need to make a good conservative guess about which segment to
-    // fetch
-    if (this.fetchAtBuffer_) {
-      // Find the segment containing the end of the buffer
-      let mediaSourceInfo = getMediaInfoForTime(playlist,
-                                                lastBufferedEnd,
-                                                syncPoint.segmentIndex,
-                                                syncPoint.time);
-
-      mediaIndex = mediaSourceInfo.mediaIndex;
-      startOfSegment = mediaSourceInfo.startTime;
-    } else {
-      // Find the segment containing currentTime
-      let mediaSourceInfo = getMediaInfoForTime(playlist,
-                                                currentTime,
-                                                syncPoint.segmentIndex,
-                                                syncPoint.time);
-
-      mediaIndex = mediaSourceInfo.mediaIndex;
-      startOfSegment = mediaSourceInfo.startTime;
-    }
-    this.logger_('getMediaIndexForTime',
-      'mediaIndex:', mediaIndex,
-      'startOfSegment:', startOfSegment);
-
-    return this.generateSegmentInfo_(playlist, mediaIndex, startOfSegment, false);
-  }
-
-  /**
-   * The segment loader has no recourse except to fetch a segment in the
-   * current playlist and use the internal timestamps in that segment to
-   * generate a syncPoint. This function returns a good candidate index
-   * for that process.
-   *
-   * @param {Object} playlist - the playlist object to look for a
-   * @returns {Number} An index of a segment from the playlist to load
-   */
-  getSyncSegmentCandidate_(playlist) {
-    if (this.currentTimeline_ === -1) {
-      return 0;
-    }
-
-    let segmentIndexArray = playlist.segments
-      .map((s, i) => {
-        return {
-          timeline: s.timeline,
-          segmentIndex: i
-        };
-      }).filter(s => s.timeline === this.currentTimeline_);
-
-    if (segmentIndexArray.length) {
-      return segmentIndexArray[Math.min(segmentIndexArray.length - 1, 1)].segmentIndex;
-    }
-
-    return Math.max(playlist.segments.length - 1, 0);
-  }
-
-  generateSegmentInfo_(playlist, mediaIndex, startOfSegment, isSyncRequest) {
-    if (mediaIndex < 0 || mediaIndex >= playlist.segments.length) {
-      return null;
-    }
-
-    let segment = playlist.segments[mediaIndex];
-
-    return {
-      requestId: 'segment-loader-' + Math.random(),
-      // resolve the segment URL relative to the playlist
-      uri: segment.resolvedUri,
-      // the segment's mediaIndex at the time it was requested
-      mediaIndex,
-      // whether or not to update the SegmentLoader's state with this
-      // segment's mediaIndex
-      isSyncRequest,
-      startOfSegment,
-      // the segment's playlist
-      playlist,
-      // unencrypted bytes of the segment
-      bytes: null,
-      // when a key is defined for this segment, the encrypted bytes
-      encryptedBytes: null,
-      // The target timestampOffset for this segment when we append it
-      // to the source buffer
-      timestampOffset: null,
-      // The timeline that the segment is in
-      timeline: segment.timeline,
-      // The expected duration of the segment in seconds
-      duration: segment.duration,
-      // retain the segment in case the playlist updates while doing an async process
-      segment
-    };
   }
 
   /**
@@ -680,9 +194,9 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
    * empty or not.
    *
    * @param {Object} segmentInfo
-   *        a segment request object that describes the segment to load
+   *        a segment info object that describes the current segment
    * @return {Object}
-   *         a segment request object that describes the segment to load
+   *         a segment info object that describes the current segment
    */
   skipEmptySegments_(segmentInfo) {
     while (segmentInfo && segmentInfo.segment.empty) {
@@ -692,198 +206,6 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
                                               segmentInfo.isSyncRequest);
     }
     return segmentInfo;
-  }
-
-  /**
-   * load a specific segment from a request into the buffer
-   *
-   * @private
-   */
-  loadSegment_(segmentInfo) {
-    this.state = 'WAITING';
-    this.pendingSegment_ = segmentInfo;
-    this.trimBackBuffer_(segmentInfo);
-
-    segmentInfo.abortRequests = mediaSegmentRequest(this.hls_.xhr,
-      this.xhrOptions_,
-      this.decrypter_,
-      this.createSimplifiedSegmentObj_(segmentInfo),
-      // progress callback
-      (event, segment) => {
-        if (!this.pendingSegment_ || segment.requestId !== this.pendingSegment_.requestId) {
-          return;
-        }
-        // TODO: Use progress-based bandwidth to early abort low-bandwidth situations
-        this.trigger('progress');
-      },
-      this.segmentRequestFinished_.bind(this));
-  }
-
-  /**
-   * trim the back buffer so that we don't have too much data
-   * in the source buffer
-   *
-   * @private
-   *
-   * @param {Object} segmentInfo - the current segment
-   */
-  trimBackBuffer_(segmentInfo) {
-    const seekable = this.seekable_();
-    const currentTime = this.currentTime_();
-    let removeToTime = 0;
-
-    // Chrome has a hard limit of 150mb of
-    // buffer and a very conservative "garbage collector"
-    // We manually clear out the old buffer to ensure
-    // we don't trigger the QuotaExceeded error
-    // on the source buffer during subsequent appends
-
-    // If we have a seekable range use that as the limit for what can be removed safely
-    // otherwise remove anything older than 1 minute before the current play head
-    if (seekable.length &&
-        seekable.start(0) > 0 &&
-        seekable.start(0) < currentTime) {
-      removeToTime = seekable.start(0);
-    } else {
-      removeToTime = currentTime - 60;
-    }
-
-    if (removeToTime > 0) {
-      this.remove(0, removeToTime);
-    }
-  }
-
-  /**
-   * created a simplified copy of the segment object with just the
-   * information necessary to perform the XHR and decryption
-   *
-   * @private
-   *
-   * @param {Object} segmentInfo - the current segment
-   * @returns {Object} a simplified segment object copy
-   */
-  createSimplifiedSegmentObj_(segmentInfo) {
-    const segment = segmentInfo.segment;
-    const simpleSegment = {
-      resolvedUri: segment.resolvedUri,
-      byterange: segment.byterange,
-      requestId: segmentInfo.requestId
-    };
-
-    if (segment.key) {
-      // if the media sequence is greater than 2^32, the IV will be incorrect
-      // assuming 10s segments, that would be about 1300 years
-      const iv = segment.key.iv || new Uint32Array([
-        0, 0, 0, segmentInfo.mediaIndex + segmentInfo.playlist.mediaSequence
-      ]);
-
-      simpleSegment.key = {
-        resolvedUri: segment.key.resolvedUri,
-        iv
-      };
-    }
-
-    if (segment.map) {
-      simpleSegment.map = this.initSegment(segment.map);
-    }
-
-    return simpleSegment;
-  }
-
-  /**
-   * Handle the callback from the segmentRequest function and set the
-   * associated SegmentLoader state and errors if necessary
-   *
-   * @private
-   */
-  segmentRequestFinished_(error, simpleSegment) {
-    // every request counts as a media request even if it has been aborted
-    // or canceled due to a timeout
-    this.mediaRequests += 1;
-
-    if (simpleSegment.stats) {
-      this.mediaBytesTransferred += simpleSegment.stats.bytesReceived;
-      this.mediaTransferDuration += simpleSegment.stats.roundTripTime;
-    }
-
-    // The request was aborted and the SegmentLoader has already been reset
-    if (!this.pendingSegment_) {
-      this.mediaRequestsAborted += 1;
-      return;
-    }
-
-    // the request was aborted and the SegmentLoader has already started
-    // another request. this can happen when the timeout for an aborted
-    // request triggers due to a limitation in the XHR library
-    // do not count this as any sort of request or we risk double-counting
-    if (simpleSegment.requestId !== this.pendingSegment_.requestId) {
-      return;
-    }
-
-    // an error occurred from the active pendingSegment_ so reset everything
-    if (error) {
-      this.pendingSegment_ = null;
-
-      // the requests were aborted just record the aborted stat and exit
-      // this is not a true error condition and nothing corrective needs
-      // to be done
-      if (error.code === REQUEST_ERRORS.ABORTED) {
-        this.mediaRequestsAborted += 1;
-        return;
-      }
-
-      this.state = 'READY';
-      this.pause();
-
-      // the error is really just that at least one of the requests timed-out
-      // set the bandwidth to a very low value and trigger an ABR switch to
-      // take emergency action
-      if (error.code === REQUEST_ERRORS.TIMEOUT) {
-        this.mediaRequestsTimedout += 1;
-        this.bandwidth = 1;
-        this.roundTrip = NaN;
-        this.trigger('bandwidthupdate');
-        return;
-      }
-
-      // if control-flow has arrived here, then the error is real
-      // emit an error event to blacklist the current playlist
-      this.mediaRequestsErrored += 1;
-      this.error(error);
-      this.trigger('error');
-      return;
-    }
-
-    // the response was a success so set any bandwidth stats the request
-    // generated for ABR purposes
-    this.bandwidth = simpleSegment.stats.bandwidth;
-    this.roundTrip = simpleSegment.stats.roundTripTime;
-
-    // if this request included an initialization segment, save that data
-    // to the initSegment cache
-    if (simpleSegment.map) {
-      simpleSegment.map = this.initSegment(simpleSegment.map, true);
-    }
-
-    this.processSegmentResponse_(simpleSegment);
-  }
-
-  /**
-   * Move any important data from the simplified segment object
-   * back to the real segment object for future phases
-   *
-   * @private
-   */
-  processSegmentResponse_(simpleSegment) {
-    const segmentInfo = this.pendingSegment_;
-
-    segmentInfo.bytes = simpleSegment.bytes;
-    if (simpleSegment.map) {
-      segmentInfo.segment.map.bytes = simpleSegment.map.bytes;
-    }
-
-    segmentInfo.endOfAllRequests = simpleSegment.endOfAllRequests;
-    this.handleSegment_();
   }
 
   /**
@@ -961,6 +283,13 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
     this.handleUpdateEnd_();
   }
 
+  /**
+   * Uses the WebVTT parser to parse the segment response
+   *
+   * @param {Object} segmentInfo
+   *        a segment info object that describes the current segment
+   * @private
+   */
   parseVTTCues_(segmentInfo) {
     let decoder;
     let decodeBytesToString = false;
@@ -1005,6 +334,19 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
     parser.flush();
   }
 
+  /**
+   * Updates the start and end times of any cues parsed by the WebVTT parser using
+   * the information parsed from the X-TIMESTAMP-MAP header and a TS to media time mapping
+   * from the SyncController
+   *
+   * @param {Object} segmentInfo
+   *        a segment info object that describes the current segment
+   * @param {Object} mappingObj
+   *        object containing a mapping from TS to media time
+   * @param {Object} playlist
+   *        the playlist object containing the segment
+   * @private
+   */
   updateTimeMapping_(segmentInfo, mappingObj, playlist) {
     const segment = segmentInfo.segment;
 
@@ -1042,96 +384,4 @@ export default class VTTSegmentLoader extends videojs.EventTarget {
       };
     }
   }
-
-  /**
-   * callback to run when appendBuffer is finished. detects if we are
-   * in a good state to do things with the data we got, or if we need
-   * to wait for more
-   *
-   * @private
-   */
-  handleUpdateEnd_() {
-    this.logger_('handleUpdateEnd_', 'segmentInfo:', this.pendingSegment_);
-
-    if (!this.pendingSegment_) {
-      this.state = 'READY';
-      if (!this.paused()) {
-        this.monitorBuffer_();
-      }
-      return;
-    }
-
-    const segmentInfo = this.pendingSegment_;
-    const segment = segmentInfo.segment;
-    const isWalkingForward = this.mediaIndex !== null;
-
-    this.pendingSegment_ = null;
-    this.recordThroughput_(segmentInfo);
-
-    this.state = 'READY';
-
-    this.mediaIndex = segmentInfo.mediaIndex;
-    this.fetchAtBuffer_ = true;
-    this.currentTimeline_ = segmentInfo.timeline;
-
-    // We must update the syncinfo to recalculate the seekable range before
-    // the following conditional otherwise it may consider this a bad "guess"
-    // and attempt to resync when the post-update seekable window and live
-    // point would mean that this was the perfect segment to fetch
-    this.trigger('syncinfoupdate');
-
-    // If we previously appended a segment that ends more than 3 targetDurations before
-    // the currentTime_ that means that our conservative guess was too conservative.
-    // In that case, reset the loader state so that we try to use any information gained
-    // from the previous request to create a new, more accurate, sync-point.
-    if (segment.end &&
-        this.currentTime_() - segment.end > segmentInfo.playlist.targetDuration * 3) {
-      this.resetEverything();
-      return;
-    }
-
-    // Don't do a rendition switch unless we have enough time to get a sync segment
-    // and conservatively guess
-    if (isWalkingForward) {
-      this.trigger('bandwidthupdate');
-    }
-    this.trigger('progress');
-
-    if (!this.paused()) {
-      this.monitorBuffer_();
-    }
-  }
-
-  /**
-   * Records the current throughput of the decrypt, transmux, and append
-   * portion of the semgment pipeline. `throughput.rate` is a the cumulative
-   * moving average of the throughput. `throughput.count` is the number of
-   * data points in the average.
-   *
-   * @private
-   * @param {Object} segmentInfo the object returned by loadSegment
-   */
-  recordThroughput_(segmentInfo) {
-    const rate = this.throughput.rate;
-    // Add one to the time to ensure that we don't accidentally attempt to divide
-    // by zero in the case where the throughput is ridiculously high
-    const segmentProcessingTime =
-      Date.now() - segmentInfo.endOfAllRequests + 1;
-    // Multiply by 8000 to convert from bytes/millisecond to bits/second
-    const segmentProcessingThroughput =
-      Math.floor((segmentInfo.byteLength / segmentProcessingTime) * 8 * 1000);
-
-    // This is just a cumulative moving average calculation:
-    //   newAvg = oldAvg + (sample - oldAvg) / (sampleCount + 1)
-    this.throughput.rate +=
-      (segmentProcessingThroughput - rate) / (++this.throughput.count);
-  }
-
-  /**
-   * A debugging logger noop that is set to console.log only if debugging
-   * is enabled globally
-   *
-   * @private
-   */
-  logger_() {}
 }
